#10072: Merge asin_bw to TTNN
