+++ docs/source/ttnn/ttnn/dependencies/tt_lib.rst
@@ -828,8 +828,6 @@ Backward Operations
 
 .. autofunction:: tt_lib.tensor.polygamma_bw
 
-.. autofunction:: tt_lib.tensor.asin_bw
-
 .. autofunction:: tt_lib.tensor.asinh_bw
 
 .. autofunction:: tt_lib.tensor.cosh_bw
