+++ docs/source/ttnn/ttnn/ttnn/asin_bw.rst
@@ -1,0 +1,6 @@
+ _ttnn.asin_bw:
+
+ttnn.asin_bw
+##################
+
+.. autofunction:: ttnn.asin_bw
