+++ tests/tt_eager/python_api_testing/unit_testing/backward_ops/test_backward_asin.py
@@ -4,7 +4,7 @@
 
 import torch
 import pytest
-import tt_lib
+import ttnn
 from tests.tt_eager.python_api_testing.unit_testing.backward_ops.utility_funcs import compare_pcc, data_gen_with_range
 
 
@@ -21,7 +21,7 @@ def test_bw_asin(input_shapes, device):
     grad_data, grad_tensor = data_gen_with_range(input_shapes, -100, 100, device)
     pyt_y = torch.asin(in_data)
 
-    tt_output_tensor_on_device = tt_lib.tensor.asin_bw(grad_tensor, input_tensor)
+    tt_output_tensor_on_device = ttnn.asin_bw(grad_tensor, input_tensor)
 
     in_data.retain_grad()
