+++ ttnn/cpp/ttnn/operations/eltwise/unary_backward/device/unary_backward_op.hpp
@@ -58,6 +58,7 @@ enum class UnaryBackwardOpType {
     HARDSWISH_BW,
     TANHSHRINK_BW,
     ATANH_BW,
+    ASIN_BW
 };
 
 struct UnaryBackwardFunction{
